update to fix failing builds

+++ lib/commercetools-api/src/Models/Me/MyCartAddLineItemAction.php
@@ -64,7 +64,7 @@ interface MyCartAddLineItemAction extends MyCartUpdateAction
     public function getSku();
 
     /**
-     * @return null|float
+     * @return null|int
      */
     public function getQuantity();
 
@@ -124,9 +124,9 @@ interface MyCartAddLineItemAction extends MyCartUpdateAction
     public function setSku(?string $sku): void;
 
     /**
-     * @param ?float $quantity
+     * @param ?int $quantity
      */
-    public function setQuantity(?float $quantity): void;
+    public function setQuantity(?int $quantity): void;
 
     /**
      * @param ?ChannelResourceIdentifier $supplyChannel
