+++ lib/commercetools-api/src/Models/Me/MyCartAddLineItemActionBuilder.php
@@ -64,7 +64,7 @@ final class MyCartAddLineItemActionBuilder implements Builder
     private $sku;
 
     /**
-     * @var ?float
+     * @var ?int
      */
     private $quantity;
 
@@ -142,7 +142,7 @@ final class MyCartAddLineItemActionBuilder implements Builder
     }
 
     /**
-     * @return null|float
+     * @return null|int
      */
     public function getQuantity()
     {
@@ -256,10 +256,10 @@ final class MyCartAddLineItemActionBuilder implements Builder
     }
 
     /**
-     * @param ?float $quantity
+     * @param ?int $quantity
      * @return $this
      */
-    public function withQuantity(?float $quantity)
+    public function withQuantity(?int $quantity)
     {
         $this->quantity = $quantity;
