+++ lib/commercetools-api/src/Models/Me/MyCartAddLineItemActionModel.php
@@ -69,7 +69,7 @@ final class MyCartAddLineItemActionModel extends JsonObjectModel implements MyCa
     protected $sku;
 
     /**
-     * @var ?float
+     * @var ?int
      */
     protected $quantity;
 
@@ -109,7 +109,7 @@ final class MyCartAddLineItemActionModel extends JsonObjectModel implements MyCa
         ?string $productId = null,
         ?int $variantId = null,
         ?string $sku = null,
-        ?float $quantity = null,
+        ?int $quantity = null,
         ?ChannelResourceIdentifier $supplyChannel = null,
         ?Money $externalPrice = null,
         ?ExternalLineItemTotalPrice $externalTotalPrice = null,
@@ -254,17 +254,17 @@ final class MyCartAddLineItemActionModel extends JsonObjectModel implements MyCa
     }
 
     /**
-     * @return null|float
+     * @return null|int
      */
     public function getQuantity()
     {
         if (is_null($this->quantity)) {
-            /** @psalm-var ?float $data */
+            /** @psalm-var ?int $data */
             $data = $this->raw(self::FIELD_QUANTITY);
             if (is_null($data)) {
                 return null;
             }
-            $this->quantity = (float) $data;
+            $this->quantity = (int) $data;
         }
 
         return $this->quantity;
@@ -413,9 +413,9 @@ final class MyCartAddLineItemActionModel extends JsonObjectModel implements MyCa
     }
 
     /**
-     * @param ?float $quantity
+     * @param ?int $quantity
      */
-    public function setQuantity(?float $quantity): void
+    public function setQuantity(?int $quantity): void
     {
         $this->quantity = $quantity;
     }
