+++ lib/commercetools-api/src/Models/Me/MyCartChangeLineItemQuantityAction.php
@@ -26,7 +26,7 @@ interface MyCartChangeLineItemQuantityAction extends MyCartUpdateAction
     public function getLineItemId();
 
     /**
-     * @return null|float
+     * @return null|int
      */
     public function getQuantity();
 
@@ -46,9 +46,9 @@ interface MyCartChangeLineItemQuantityAction extends MyCartUpdateAction
     public function setLineItemId(?string $lineItemId): void;
 
     /**
-     * @param ?float $quantity
+     * @param ?int $quantity
      */
-    public function setQuantity(?float $quantity): void;
+    public function setQuantity(?int $quantity): void;
 
     /**
      * @param ?Money $externalPrice
