+++ lib/commercetools-api/src/Models/Me/MyCartChangeLineItemQuantityActionBuilder.php
@@ -30,7 +30,7 @@ final class MyCartChangeLineItemQuantityActionBuilder implements Builder
     private $lineItemId;
 
     /**
-     * @var ?float
+     * @var ?int
      */
     private $quantity;
 
@@ -53,7 +53,7 @@ final class MyCartChangeLineItemQuantityActionBuilder implements Builder
     }
 
     /**
-     * @return null|float
+     * @return null|int
      */
     public function getQuantity()
     {
@@ -88,10 +88,10 @@ final class MyCartChangeLineItemQuantityActionBuilder implements Builder
     }
 
     /**
-     * @param ?float $quantity
+     * @param ?int $quantity
      * @return $this
      */
-    public function withQuantity(?float $quantity)
+    public function withQuantity(?int $quantity)
     {
         $this->quantity = $quantity;
