+++ lib/commercetools-api/src/Models/Me/MyCartChangeLineItemQuantityActionModel.php
@@ -35,7 +35,7 @@ final class MyCartChangeLineItemQuantityActionModel extends JsonObjectModel impl
     protected $lineItemId;
 
     /**
-     * @var ?float
+     * @var ?int
      */
     protected $quantity;
 
@@ -55,7 +55,7 @@ final class MyCartChangeLineItemQuantityActionModel extends JsonObjectModel impl
      */
     public function __construct(
         ?string $lineItemId = null,
-        ?float $quantity = null,
+        ?int $quantity = null,
         ?Money $externalPrice = null,
         ?ExternalLineItemTotalPrice $externalTotalPrice = null
     ) {
@@ -101,17 +101,17 @@ final class MyCartChangeLineItemQuantityActionModel extends JsonObjectModel impl
     }
 
     /**
-     * @return null|float
+     * @return null|int
      */
     public function getQuantity()
     {
         if (is_null($this->quantity)) {
-            /** @psalm-var ?float $data */
+            /** @psalm-var ?int $data */
             $data = $this->raw(self::FIELD_QUANTITY);
             if (is_null($data)) {
                 return null;
             }
-            $this->quantity = (float) $data;
+            $this->quantity = (int) $data;
         }
 
         return $this->quantity;
@@ -163,9 +163,9 @@ final class MyCartChangeLineItemQuantityActionModel extends JsonObjectModel impl
     }
 
     /**
-     * @param ?float $quantity
+     * @param ?int $quantity
      */
-    public function setQuantity(?float $quantity): void
+    public function setQuantity(?int $quantity): void
     {
         $this->quantity = $quantity;
     }
