+++ lib/commercetools-api/src/Models/Me/MyCartRemoveLineItemActionBuilder.php
@@ -32,7 +32,7 @@ final class MyCartRemoveLineItemActionBuilder implements Builder
     private $lineItemId;
 
     /**
-     * @var ?float
+     * @var ?int
      */
     private $quantity;
 
@@ -60,7 +60,7 @@ final class MyCartRemoveLineItemActionBuilder implements Builder
     }
 
     /**
-     * @return null|float
+     * @return null|int
      */
     public function getQuantity()
     {
@@ -103,10 +103,10 @@ final class MyCartRemoveLineItemActionBuilder implements Builder
     }
 
     /**
-     * @param ?float $quantity
+     * @param ?int $quantity
      * @return $this
      */
-    public function withQuantity(?float $quantity)
+    public function withQuantity(?int $quantity)
     {
         $this->quantity = $quantity;
