+++ lib/commercetools-api/src/Models/Me/MyCartRemoveLineItemActionModel.php
@@ -37,7 +37,7 @@ final class MyCartRemoveLineItemActionModel extends JsonObjectModel implements M
     protected $lineItemId;
 
     /**
-     * @var ?float
+     * @var ?int
      */
     protected $quantity;
 
@@ -62,7 +62,7 @@ final class MyCartRemoveLineItemActionModel extends JsonObjectModel implements M
      */
     public function __construct(
         ?string $lineItemId = null,
-        ?float $quantity = null,
+        ?int $quantity = null,
         ?Money $externalPrice = null,
         ?ExternalLineItemTotalPrice $externalTotalPrice = null,
         ?ItemShippingDetailsDraft $shippingDetailsToRemove = null
@@ -110,17 +110,17 @@ final class MyCartRemoveLineItemActionModel extends JsonObjectModel implements M
     }
 
     /**
-     * @return null|float
+     * @return null|int
      */
     public function getQuantity()
     {
         if (is_null($this->quantity)) {
-            /** @psalm-var ?float $data */
+            /** @psalm-var ?int $data */
             $data = $this->raw(self::FIELD_QUANTITY);
             if (is_null($data)) {
                 return null;
             }
-            $this->quantity = (float) $data;
+            $this->quantity = (int) $data;
         }
 
         return $this->quantity;
@@ -190,9 +190,9 @@ final class MyCartRemoveLineItemActionModel extends JsonObjectModel implements M
     }
 
     /**
-     * @param ?float $quantity
+     * @param ?int $quantity
      */
-    public function setQuantity(?float $quantity): void
+    public function setQuantity(?int $quantity): void
     {
         $this->quantity = $quantity;
     }
